feat: restrict access for update and delete reports

+++ app/Http/Controllers/ReportController.php
@@ -1,18 +1,13 @@
 <?php
 
-declare(strict_types=1);
+declare (strict_types = 1);
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App\Models\Conference;
-use App\Service\CountryService;
-use App\Models\User;
-use App\Models\Report;
-use App\Http\Requests\CreateConferenceRequest;
 use App\Http\Requests\CreateReportRequest;
-use App\Http\Requests\UpdateConferenceRequest;
-use Auth, Redirect, Validator;
+use App\Models\Report;
+use App\Models\User;
+use Auth;
 
 class ReportController extends Controller
 {
@@ -20,13 +15,13 @@ class ReportController extends Controller
     {
         return response()->json(Report::All()->where('conference_id', $id));
     }
-    
+
     public function show(int $conference_id, int $report_id)
     {
         return response()->json(Report::All()->where('conference_id', $conference_id)->where('id', $report_id));
     }
 
-    public function store(int $id, CreateReportRequest $request)
+    public function store(CreateReportRequest $request, int $id)
     {
         $data = $request->validated();
         $data['conference_id'] = $id;
@@ -35,15 +30,22 @@ class ReportController extends Controller
     }
 
     public function update(CreateReportRequest $request, int $conference_id, int $report_id)
-    {   
-        $data = $request->validated();
-        $data['conference_id'] = $conference_id;
-        $data['user_id'] = Auth::user()->id;
-        Report::whereId($report_id)->update($data);
+    {
+        $rep = Report::findOrFail($report_id);
+        if ($rep->user_id === Auth::user()->id) {
+            $data = $request->validated();
+            $data['conference_id'] = $conference_id;
+            $data['user_id'] = Auth::user()->id;
+            Report::whereId($report_id)->update($data);
+        }
     }
 
     public function destroy(int $id)
     {
-        Report::findOrFail($id)->delete();
+        $rep = Report::findOrFail($report_id);
+        if ($rep->user_id === Auth::user()->id) {
+            Report::findOrFail($id)->delete();
+        }
     }
+
 }
